refactor(prover): move auto_storage and templates inline tests to tests/
Continue relocating #[cfg(test)] mod tests blocks out of prover source
files into prover/src/tests/.

- auto_storage.rs tests -> tests/auto_storage_tests.rs, gated on the
  disk-spill feature to match the auto_storage module
- constraints/templates.rs test -> tests/templates_tests.rs
- select_storage_mode and the GOLDILOCKS_BYTES / CUBIC_EXT_BYTES /
  KECCAK_NODE_BYTES consts bumped to pub(crate) for the relocated tests
